+++ VectorPrinter.h
@@ -1,14 +1,14 @@
 #pragma once
 
-#include "CStringOp.h"
+#include "StringOp.h"
 
 namespace VectorPrinter {
 	template<typename T>
-	void print(const std::vector<T>& vt, size_t level);
+	void print(const std::vector<T>& vt, int level);
 };
 
 template<typename T>
-inline void VectorPrinter::print(const std::vector<T>& vt, size_t level)
+inline void VectorPrinter::print(const std::vector<T>& vt, int level)
 {
 	std::string prefix = CStringOp::repeat("\t", level);
 	std::cout << prefix << "{" << std::endl;
